requirements.txt
dill==0.3.4
hummingbird-ml==0.4.0
joblib==1.0.1
numpy==1.21.1
onnx==1.9.0
onnxconverter-common==1.8.1
protobuf==3.17.3
psutil==5.8.0
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
threadpoolctl==2.2.0
torch==1.9.0
typing-extensions==3.10.0.0
